toolshed/requirements.txt
Django
django-filter
djangorestframework-jwt
djangorestframework
requests
